refactored components and component routes
